Updated test script to fix problem with changing elections. Removed unused tests.
