+++ package.json
@@ -56,9 +56,7 @@
     "ballotTest-Android": "./node_modules/.bin/wdio ./tests/browserstack/ballotAndroid.wdio.config.js",
     "ballotTest-Browser": "./node_modules/.bin/wdio ./tests/browserstack/ballotBrowser.wdio.config.js",
     "ballotTest-iOS": "./node_modules/.bin/wdio ./tests/browserstack/ballotIos.wdio.config.js",
-    "marketingTest-Android": "./node_modules/.bin/wdio ./tests/browserstack/marketingAndroid.wdio.config.js",
     "marketingTest-Browser": "./node_modules/.bin/wdio ./tests/browserstack/marketingBrowser.wdio.config.js",
-    "marketingTest-iOS": "./node_modules/.bin/wdio ./tests/browserstack/marketingIos.wdio.config.js",
     "start": "gulp",
     "start-profile": "node --inspect-brk ./node_modules/.bin/gulp",
     "prod": "NODE_ENV=production gulp build",
