--- tests/browserstack/marketingAndroid.wdio.config.js
@@ -1,37 +1,0 @@
-const browserStackConfig = require('./browserstack.config');
-const date = new Date();
-const dateForDisplay = date.toDateString();
-const buildNameForDisplay = `${browserStackConfig.BUILD}: ${dateForDisplay}`;
-
-exports.config = {
-  user: browserStackConfig.BROWSERSTACK_USER,
-  key: browserStackConfig.BROWSERSTACK_KEY,
-  updateJob: false,
-  specs: [
-    './tests/browserstack/specs/marketingWelcomeTest.js',
-  ],
-  exclude: [],
-  capabilities: [
-    {
-      // capabilities for a cordova android test
-      name: 'marketingWelcomeTest-Android',
-      build: buildNameForDisplay,
-      device: 'Samsung Galaxy Tab S4',
-      os_version: '8.1',
-      app: browserStackConfig.BROWSERSTACK_APK_URL,
-      'browserstack.debug': true,
-      'browserstack.geoLocation': 'US',
-    },
-  ],
-  coloredLogs: true,
-  baseUrl: '',
-  waitforTimeout: 180000, // 3 minutes
-  connectionRetryTimeout: 90000,
-  connectionRetryCount: 3,
-
-  framework: 'mocha',
-  mochaOpts: {
-    ui: 'bdd',
-    timeout: 180000, // 3 minutes
-  },
-};
